This function performs the training for (co)variance testing.

This function performs the training for (co)variance testing.

.train_models_covar(
  Y1_1,
  Y2_2,
  Y1_3,
  Y2_3,
  X_1,
  X_2,
  X_3,
  Z_1,
  Z_2,
  Z_3,
  learner_y_stage1,
  learner_y_stage2,
  learner_x,
  prediction_y_stage1,
  variance = FALSE
)

Arguments

Y1_1

Measurements for Y1_1.

Y2_2

Measurements for Y2_2.

Y1_3

Measurements for Y1_3.

Y2_3

Measurements for Y2_3.

X_1

Measurements for X_1.

X_2

Measurements for X_2.

X_3

Measurements for X_3.

Z_1

Measurements for Z_1.

Z_2

Measurements for Z_2.

Z_3

Measurements for Z_3.

learner_y_stage1

Prediction model function to learn prediction of Y given X and Z in stage 1.

learner_y_stage2

Prediction model function to learn prediction of Y given X and Z in stage 2.

learner_x

Prediction model function to learn prediction of X given Z.

prediction_y_stage1

Prediction model to predict Y in stage 1.

variance

Logical variable to indicate if Y1=Y2.