ROMY-COV

This function tests for (co)variance effects between variables in Y and X, adjusting for Z.

romy_cov(
  Y,
  X,
  Z,
  Z2 = NULL,
  index_triples = NULL,
  learner_x = lm_learner,
  prediction_x = lm_predict,
  learner_y_stage1 = lm_learner,
  prediction_y_stage1 = lm_predict,
  learner_y_stage2 = lm_learner,
  prediction_y_stage2 = lm_predict,
  K = 5,
  parallel = FALSE,
  BPPARAM = NULL
)

Arguments

Y

A matrix containing the measurements for Y.

X

A matrix containing the measurements for X.

Z

A matrix containing the measurements for Z, for adjusting Z. No missing values allowed.

Z2

A matrix containing the measurements for Z2, for adjusting X. No missing values allowed, default is NULL. In this case, Z2=Z.

index_triples

A dataframe with three columns containing the indices in X and Y to be tested, respectively. Default is NULL, leading to testing all possible triples.

learner_x

Prediction model function to learn prediction of X given Z. Defaul is linear regression.

prediction_x

Prediction model to predict X given Z. Defaul is linear regression.

learner_y_stage1

Prediction model function to learn prediction of Y given X and Z in stage 1. Default is linear regression.

prediction_y_stage1

Prediction model to predict Y given X and Z in stage 1. Default is linear regression.

learner_y_stage2

Prediction model function to learn prediction of Y in stage 2. Default is linear regression.

prediction_y_stage2

Prediction model to predict Y in stage 2. Default is linear regression.

K

Value for K for the K fold cross fitting. Default is K=5.

parallel

Logic value indicating if parallel computation should be used. Default is FALSE. If TRUE, BPPARAM needs to be initialized

BPPARAM

BPPARAM object for BiocParallel parallel computation, default is NULL.

return Dataframe containing the p-values for all tests.